frigate/endpoint/requirements.txt
requests==2.21.0
flask==0.12.4
kafka-python==1.4.6
